[ASV-1350] Reorganized architecture to have a presenter for each flavor and handle the TC and PP.
