+++ app/src/main/java/cm/aptoide/pt/account/view/LoginSignUpCredentialsFragment.java
@@ -28,8 +28,8 @@ import cm.aptoide.pt.BuildConfig;
 import cm.aptoide.pt.R;
 import cm.aptoide.pt.account.AccountAnalytics;
 import cm.aptoide.pt.orientation.ScreenOrientationManager;
-import cm.aptoide.pt.presenter.LoginSignUpCredentialsPresenter;
 import cm.aptoide.pt.presenter.LoginSignUpCredentialsView;
+import cm.aptoide.pt.presenter.LoginSignupCredentialsSpecificPresenter;
 import cm.aptoide.pt.utils.GenericDialogs;
 import cm.aptoide.pt.view.NotBottomNavigationView;
 import cm.aptoide.pt.view.rx.RxAlertDialog;
@@ -47,7 +47,7 @@ public class LoginSignUpCredentialsFragment extends GooglePlayServicesFragment
 
   private static final String USERNAME_KEY = "username_key";
   private static final String PASSWORD_KEY = "password_key";
-  @Inject LoginSignUpCredentialsPresenter presenter;
+  @Inject LoginSignupCredentialsSpecificPresenter presenter;
   @Inject ScreenOrientationManager orientationManager;
   @Inject AccountAnalytics accountAnalytics;
   @Inject @Named("marketName") String marketName;
@@ -292,7 +292,7 @@ public class LoginSignUpCredentialsFragment extends GooglePlayServicesFragment
     googleLoginButton.setVisibility(View.GONE);
   }
 
-  @Override public boolean tryCloseLoginBottomSheet() {
+  @Override public boolean tryCloseLoginBottomSheet(boolean shouldShowTCandPP) {
     if (credentialsEditTextsArea.getVisibility() == View.VISIBLE) {
       bottomSheetBehavior.setPeekHeight(originalHeight);
       credentialsEditTextsArea.setVisibility(View.GONE);
@@ -300,8 +300,10 @@ public class LoginSignUpCredentialsFragment extends GooglePlayServicesFragment
       loginArea.setVisibility(View.GONE);
       signUpArea.setVisibility(View.GONE);
       separator.setVisibility(View.VISIBLE);
-      termsConditionCheckBox.setVisibility(View.VISIBLE);
-      termsAndConditions.setVisibility(View.VISIBLE);
+      if (shouldShowTCandPP) {
+        termsConditionCheckBox.setVisibility(View.VISIBLE);
+        termsAndConditions.setVisibility(View.VISIBLE);
+      }
       return true;
     }
     return false;
@@ -383,8 +385,6 @@ public class LoginSignUpCredentialsFragment extends GooglePlayServicesFragment
     credentialsEditTextsArea = view.findViewById(R.id.credentials_edit_texts);
     signUpSelectionButton = (Button) view.findViewById(R.id.show_join_aptoide_area);
     loginSelectionButton = (Button) view.findViewById(R.id.show_login_with_aptoide_area);
-    termsConditionCheckBox = (CheckBox) view.findViewById(R.id.tc_checkbox);
-    checkboxDrawable = CompoundButtonCompat.getButtonDrawable(termsConditionCheckBox);
 
     if ("vanilla".equalsIgnoreCase(BuildConfig.FLAVOR_product)) {
       buttonSignUp.setText(String.format(getString(R.string.onboarding_button_join_us)));
@@ -394,6 +394,39 @@ public class LoginSignUpCredentialsFragment extends GooglePlayServicesFragment
     loginArea = view.findViewById(R.id.login_button_area);
     signUpArea = view.findViewById(R.id.sign_up_button_area);
     separator = view.findViewById(R.id.separator);
+
+    facebookEmailRequiredDialog = new RxAlertDialog.Builder(getContext()).setMessage(
+        R.string.facebook_email_permission_regected_message)
+        .setPositiveButton(R.string.facebook_grant_permission_button)
+        .setNegativeButton(android.R.string.cancel)
+        .build();
+
+    termsConditionCheckBox = (CheckBox) view.findViewById(R.id.tc_checkbox);
+    termsAndConditions = (TextView) view.findViewById(R.id.terms_and_conditions);
+
+    progressDialog = GenericDialogs.createGenericPleaseWaitDialog(getContext());
+
+    try {
+      bottomSheetBehavior = BottomSheetBehavior.from(view.getRootView()
+          .findViewById(R.id.login_signup_layout));
+    } catch (IllegalArgumentException ex) {
+      // this happens because in landscape the R.id.login_signup_layout is not
+      // a child of CoordinatorLayout
+    }
+
+    originalHeight = bottomSheetBehavior.getPeekHeight();
+    attachPresenter(presenter);
+    registerClickHandler(presenter);
+  }
+
+  public void hideTCandPP() {
+    termsConditionCheckBox.setVisibility(View.GONE);
+    termsAndConditions.setVisibility(View.GONE);
+  }
+
+  public void showTCandPP() {
+    checkboxDrawable = CompoundButtonCompat.getButtonDrawable(termsConditionCheckBox);
+    termsConditionCheckBox.setVisibility(View.VISIBLE);
 
     ClickableSpan termsAndConditionsClickListener = new ClickableSpan() {
       @Override public void onClick(View view) {
@@ -427,29 +460,9 @@ public class LoginSignUpCredentialsFragment extends GooglePlayServicesFragment
         privacyAndTerms.indexOf(privacyPolicyPlaceHolder) + privacyPolicyPlaceHolder.length(),
         Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
 
-    termsAndConditions = (TextView) view.findViewById(R.id.terms_and_conditions);
     termsAndConditions.setText(privacyAndTermsSpan);
     termsAndConditions.setMovementMethod(LinkMovementMethod.getInstance());
-
-    facebookEmailRequiredDialog = new RxAlertDialog.Builder(getContext()).setMessage(
-        R.string.facebook_email_permission_regected_message)
-        .setPositiveButton(R.string.facebook_grant_permission_button)
-        .setNegativeButton(android.R.string.cancel)
-        .build();
-
-    progressDialog = GenericDialogs.createGenericPleaseWaitDialog(getContext());
-
-    try {
-      bottomSheetBehavior = BottomSheetBehavior.from(view.getRootView()
-          .findViewById(R.id.login_signup_layout));
-    } catch (IllegalArgumentException ex) {
-      // this happens because in landscape the R.id.login_signup_layout is not
-      // a child of CoordinatorLayout
-    }
-
-    originalHeight = bottomSheetBehavior.getPeekHeight();
-    attachPresenter(presenter);
-    registerClickHandler(presenter);
+    termsAndConditions.setVisibility(View.VISIBLE);
   }
 
   @Override public void onDestroyView() {
