+++ app/src/main/java/cm/aptoide/pt/presenter/LoginSignUpCredentialsPresenter.java
@@ -19,7 +19,8 @@ import java.util.Collection;
 import rx.Observable;
 import rx.android.schedulers.AndroidSchedulers;
 
-public class LoginSignUpCredentialsPresenter implements Presenter, BackButton.ClickHandler {
+public abstract class LoginSignUpCredentialsPresenter
+    implements Presenter, BackButton.ClickHandler {
 
   private static final int RESOLVE_GOOGLE_CREDENTIALS_REQUEST_CODE = 2;
   private final LoginSignUpCredentialsView view;
@@ -55,9 +56,6 @@ public class LoginSignUpCredentialsPresenter implements Presenter, BackButton.Cl
 
     handleAptoideLoginEvent();
 
-    handleClickOnTermsAndConditions();
-    handleClickOnPrivacyPolicy();
-
     handleGoogleSignUpEvent();
     handleGoogleSignUpResult();
 
@@ -66,31 +64,9 @@ public class LoginSignUpCredentialsPresenter implements Presenter, BackButton.Cl
     handleFacebookSignUpWithRequiredPermissionsEvent();
 
     handleAptoideShowLoginEvent();
-    handleAptoideShowSignUpEvent();
-    handleAptoideSignUpEvent();
     handleAccountStatusChangeWhileShowingView();
     handleForgotPasswordClick();
     handleTogglePasswordVisibility();
-  }
-
-  private void handleClickOnTermsAndConditions() {
-    view.getLifecycleEvent()
-        .filter(event -> event.equals(View.LifecycleEvent.CREATE))
-        .flatMap(__ -> view.termsAndConditionsClickEvent())
-        .doOnNext(__ -> accountNavigator.navigateToTermsAndConditions())
-        .compose(view.bindUntilEvent(View.LifecycleEvent.DESTROY))
-        .subscribe(__ -> {
-        }, err -> crashReport.log(err));
-  }
-
-  private void handleClickOnPrivacyPolicy() {
-    view.getLifecycleEvent()
-        .filter(event -> event.equals(View.LifecycleEvent.CREATE))
-        .flatMap(__ -> view.privacyPolicyClickEvent())
-        .doOnNext(__ -> accountNavigator.navigateToPrivacyPolicy())
-        .compose(view.bindUntilEvent(View.LifecycleEvent.DESTROY))
-        .subscribe(__ -> {
-        }, err -> crashReport.log(err));
   }
 
   private void handleTogglePasswordVisibility() {
@@ -140,58 +116,10 @@ public class LoginSignUpCredentialsPresenter implements Presenter, BackButton.Cl
         .subscribe();
   }
 
-  private void handleAptoideSignUpEvent() {
-    view.getLifecycleEvent()
-        .filter(event -> event.equals(View.LifecycleEvent.CREATE))
-        .flatMap(__ -> view.aptoideSignUpEvent()
-            .doOnNext(credentials -> showNotCheckedMessage(credentials.isChecked()))
-            .filter(AptoideCredentials::isChecked)
-            .doOnNext(click -> {
-              view.hideKeyboard();
-              view.showLoading();
-              lockScreenRotation();
-              accountAnalytics.sendAptoideSignUpButtonPressed();
-            })
-            .flatMapCompletable(
-                credentials -> accountManager.signUp(AptoideAccountManager.APTOIDE_SIGN_UP_TYPE,
-                    credentials)
-                    .observeOn(AndroidSchedulers.mainThread())
-                    .doOnCompleted(() -> {
-                      accountAnalytics.loginSuccess();
-                      navigateToCreateProfile();
-                      unlockScreenRotation();
-                      view.hideLoading();
-                    })
-                    .doOnError(throwable -> {
-                      accountAnalytics.sendSignUpErrorEvent(AccountAnalytics.LoginMethod.APTOIDE,
-                          throwable);
-                      view.showError(errorMapper.map(throwable));
-                      crashReport.log(throwable);
-                      unlockScreenRotation();
-                      view.hideLoading();
-                    }))
-            .retry())
-        .compose(view.bindUntilEvent(View.LifecycleEvent.DESTROY))
-        .subscribe();
-  }
-
   private void handleAptoideShowLoginEvent() {
     view.getLifecycleEvent()
         .filter(event -> event.equals(View.LifecycleEvent.CREATE))
         .flatMap(__ -> aptoideShowLoginClick())
-        .compose(view.bindUntilEvent(View.LifecycleEvent.DESTROY))
-        .subscribe(__ -> {
-        }, err -> {
-          view.hideLoading();
-          view.showError(errorMapper.map(err));
-          crashReport.log(err);
-        });
-  }
-
-  private void handleAptoideShowSignUpEvent() {
-    view.getLifecycleEvent()
-        .filter(event -> event.equals(View.LifecycleEvent.CREATE))
-        .flatMap(__ -> showAptoideSignUpEvent())
         .compose(view.bindUntilEvent(View.LifecycleEvent.DESTROY))
         .subscribe(__ -> {
         }, err -> {
@@ -335,13 +263,6 @@ public class LoginSignUpCredentialsPresenter implements Presenter, BackButton.Cl
         .doOnNext(__ -> view.showAptoideLoginArea());
   }
 
-  private Observable<Boolean> showAptoideSignUpEvent() {
-    return view.showAptoideSignUpAreaClick()
-        .doOnNext(this::showNotCheckedMessage)
-        .filter(event -> event)
-        .doOnNext(__ -> view.showAptoideSignUpArea());
-  }
-
   private Observable<Void> forgotPasswordSelection() {
     return view.forgotPasswordClick()
         .doOnNext(selection -> accountNavigator.navigateToRecoverPasswordView());
@@ -384,19 +305,15 @@ public class LoginSignUpCredentialsPresenter implements Presenter, BackButton.Cl
     }
   }
 
-  @Override public boolean handle() {
-    return view.tryCloseLoginBottomSheet();
-  }
-
-  private void lockScreenRotation() {
+  void lockScreenRotation() {
     view.lockScreenRotation();
   }
 
-  private void unlockScreenRotation() {
+  void unlockScreenRotation() {
     view.unlockScreenRotation();
   }
 
-  private void navigateToCreateProfile() {
+  void navigateToCreateProfile() {
     accountNavigator.navigateToCreateProfileView();
   }
 
@@ -406,11 +323,5 @@ public class LoginSignUpCredentialsPresenter implements Presenter, BackButton.Cl
 
   private void navigateBack() {
     accountNavigator.popView();
-  }
-
-  private void showNotCheckedMessage(boolean checked) {
-    if (!checked) {
-      view.showTermsConditionError();
-    }
   }
 }
