+++ app/src/main/java/cm/aptoide/pt/presenter/LoginSignUpCredentialsView.java
@@ -64,7 +64,7 @@ public interface LoginSignUpCredentialsView extends GooglePlayServicesView {
 
   void hideGoogleLogin();
 
-  boolean tryCloseLoginBottomSheet();
+  boolean tryCloseLoginBottomSheet(boolean shouldShowTCandPP);
 
   boolean isPasswordVisible();
 
@@ -73,4 +73,8 @@ public interface LoginSignUpCredentialsView extends GooglePlayServicesView {
   void lockScreenRotation();
 
   void unlockScreenRotation();
+
+  void hideTCandPP();
+
+  void showTCandPP();
 }
