+++ app/src/main/res/layout/fragment_login_sign_up_credentials.xml
@@ -95,6 +95,7 @@
         android:layout_marginTop="8dp"
         android:paddingBottom="8dp"
         android:paddingTop="14dp"
+        android:visibility="gone"
         />
     <TextView
         android:id="@+id/terms_and_conditions"
@@ -108,9 +109,10 @@
         android:gravity="start"
         android:paddingBottom="8dp"
         android:paddingTop="14dp"
-        android:text="@string/accept_terms_signup"
+        tools:text="@string/accept_terms_signup"
         android:textColor="@color/black_text"
         style="@style/Aptoide.TextView.Regular.XS"
+        android:visibility="gone"
         />
   </LinearLayout>
